general fixes

+++ client/include/client_global.h
@@ -111,7 +111,7 @@ namespace global {
   inline bool chat_open;
   inline bool inv_open;
 
-  inline double delta_time;
+  inline float delta_time;
 
   inline Vector3 last_movement;
   inline Vector3 next_movement;
